+++ src/create-store-internals.js
@@ -1,5 +1,6 @@
 import memoizerific from 'memoizerific';
 import { createDraft, finishDraft, isDraft } from 'immer-peasy';
+import isPlainObject from 'is-plain-object';
 import {
   actionOnSymbol,
   actionSymbol,
@@ -9,37 +10,8 @@ import {
   thunkOnSymbol,
   thunkSymbol,
 } from './constants';
-import { isStateObject, get, set, isPromise } from './lib';
-
-const newify = (currentPath, currentState, finalValue) => {
-  if (currentPath.length === 0) {
-    return finalValue;
-  }
-  const newState = { ...currentState };
-  const key = currentPath[0];
-  if (currentPath.length === 1) {
-    newState[key] = finalValue;
-  } else {
-    newState[key] = newify(currentPath.slice(1), newState[key], finalValue);
-  }
-  return newState;
-};
-
-const noopStorage = {
-  getItem: () => undefined,
-  setItem: () => undefined,
-  removeItem: () => undefined,
-};
-
-const localStorage =
-  typeof window !== 'undefined' && typeof window.localStorage !== 'undefined'
-    ? window.localStorage
-    : noopStorage;
-
-const sessionStorage =
-  typeof window !== 'undefined' && typeof window.sessionStorage !== 'undefined'
-    ? window.sessionStorage
-    : noopStorage;
+import { get, set, newify } from './lib';
+import { createStorageWrapper } from './storage';
 
 export default function createStoreInternals({
   disableImmer,
@@ -65,18 +37,17 @@ export default function createStoreInternals({
         return isDraft(result) ? finishDraft(result) : result;
       }
       return finishDraft(draft);
-    } else {
-      const parentPath = path.slice(0, path.length - 1);
-      const draft = createDraft(state);
-      const parent = get(parentPath, state);
-      const current = get(path, draft);
-      const result = fn(current);
-
-      if (result) {
-        parent[path[path.length - 1]] = result;
-      }
-      return finishDraft(draft);
     }
+    const parentPath = path.slice(0, path.length - 1);
+    const draft = createDraft(state);
+    const parent = get(parentPath, state);
+    const current = get(path, draft);
+    const result = fn(current);
+
+    if (result) {
+      parent[path[path.length - 1]] = result;
+    }
+    return finishDraft(draft);
   }
 
   const defaultState = initialState;
@@ -114,84 +85,16 @@ export default function createStoreInternals({
       };
 
       if (key === persistSymbol) {
-        let config = value || {};
-        const {
-          blacklist = [],
-          mergeStrategy = 'merge',
-          transformers = [],
-          whitelist = [],
-        } = config;
-        let storage;
-        if (config.storage == null) {
-          storage = sessionStorage;
-        } else if (typeof config.storage === 'string') {
-          if (config.storage === 'localStorage') {
-            storage = localStorage;
-          } else if (config.storage === 'sessionStorage') {
-            storage = sessionStorage;
-          } else {
-            if (process.env.NODE_ENV === 'development') {
-              console.warn(
-                `Invalid storage provider specified for Easy Peasy persist: ${config.storage}\nValid values include "localStorage", "sessionStorage" or a custom storage engine.`,
-              );
-            }
-            storage = noopStorage;
-          }
-        } else {
-          storage = config.storage;
-        }
-
-        const outTransformers = transformers.reverse();
-
-        const serialize = (data, key) => {
-          const simpleKey = key.substr(key.indexOf('@') + 1);
-          const transformed = transformers.reduce((acc, cur) => {
-            return cur.in(acc, simpleKey);
-          }, data);
-          return storage === localStorage || storage === sessionStorage
-            ? JSON.stringify({ data: transformed })
-            : transformed;
-        };
-        const deserialize = (data, key) => {
-          const simpleKey = key.substr(key.indexOf('@') + 1);
-          const result =
-            storage === localStorage || storage === sessionStorage
-              ? JSON.parse(data).data
-              : data;
-          return outTransformers.reduce((acc, cur) => {
-            return cur.out(acc, simpleKey);
-          }, result);
-        };
-        const storageWrapper = {
-          getItem: key => {
-            if (isAsyncStorage) {
-              return storage.getItem(key).then(wrapped => {
-                return wrapped != null ? deserialize(wrapped, key) : undefined;
-              });
-            }
-            const wrapped = storage.getItem(key);
-            return wrapped != null ? deserialize(wrapped, key) : undefined;
-          },
-          setItem: (key, data) => {
-            return storage.setItem(key, serialize(data, key));
-          },
-          removeItem: key => {
-            return storage.removeItem(key);
-          },
-        };
-
-        const isAsyncStorage = isPromise(storage.getItem('__fooooooo__'));
+        const config = value || {};
         persistenceConfig.push({
           path: parentPath,
           config: {
-            blacklist,
-            isAsyncStorage,
-            mergeStrategy,
-            storage: storageWrapper,
-            whitelist,
+            blacklist: config.blacklist || [],
+            mergeStrategy: config.mergeStrategy || 'merge',
+            storage: createStorageWrapper(config.storage, config.transformers),
+            whitelist: config.whitelist || [],
           },
         });
-        delete current[key];
         return;
       }
 
@@ -369,7 +272,7 @@ export default function createStoreInternals({
         } else {
           handleValueAsState();
         }
-      } else if (isStateObject(value)) {
+      } else if (isPlainObject(value)) {
         const existing = get(path, defaultState);
         if (existing == null) {
           set(path, defaultState, {});
